@@ -1,80 +1,65 @@
 {
     "name": "changeweb/unifiedtransform",
-    "description": "A school management software.",
-    "keywords": ["school", "management"],
-    "license": "GPL 3",
     "type": "project",
+    "description": "The School Management App.",
+    "keywords": ["framework", "laravel"],
+    "license": "MIT",
     "require": {
-        "php": ">=7.2.5",
-        "doctrine/dbal": "^2.9",
-        "fideloper/proxy": "^4.2",
-        "lab404/laravel-impersonate": "1.6.0",
-        "laravel/cashier": "^11.0",
-        "guzzlehttp/guzzle": "^6.3",
-        "laravel/framework": "^7.0",
-        "laravel/passport": "^9.1.0",
-        "laravel/tinker": "^2.0",
-        "maatwebsite/excel": "^3.1",
-        "laravel/ui": "2.0.3",
-        "mavinoo/laravel-batch": "2.1.9",
-        "renatomarinho/laravel-page-speed": "^1.8"
+        "php": "^7.3|^8.0",
+        "fruitcake/laravel-cors": "^2.0",
+        "guzzlehttp/guzzle": "^7.0.1",
+        "laravel/framework": "^8.54",
+        "laravel/tinker": "^2.5",
+        "spatie/laravel-permission": "^4.2",
+        "stevebauman/purify": "^4.0"
     },
     "require-dev": {
-        "facade/ignition": "^2.0",
-        "barryvdh/laravel-debugbar": "^3.2",
-        "filp/whoops": "~2.0",
-        "friendsofphp/php-cs-fixer": "^2.14",
-        "fzaninotto/faker": "^1.9.1",
-        "laravel/dusk": "^6.1.0",
-        "mockery/mockery": "^1.3.1",
-        "nunomaduro/collision": "^4.1",
-        "phpunit/phpunit": "^8.5",
-        "rap2hpoutre/laravel-log-viewer": "^1.6.0"
+        "barryvdh/laravel-debugbar": "^3.6",
+        "facade/ignition": "^2.5",
+        "fakerphp/faker": "^1.9.1",
+        "laravel/sail": "^1.0.1",
+        "laravel/ui": "^3.3",
+        "mockery/mockery": "^1.4.2",
+        "nunomaduro/collision": "^5.0",
+        "phpunit/phpunit": "^9.3.3"
     },
     "autoload": {
-        "classmap": [
-            "database/seeds",
-            "database/factories"
-        ],
         "psr-4": {
-            "App\\": "app/"
+            "App\\": "app/",
+            "Database\\Factories\\": "database/factories/",
+            "Database\\Seeders\\": "database/seeders/"
         }
     },
     "autoload-dev": {
         "psr-4": {
             "Tests\\": "tests/"
-        },
-        "files": [
-            "tests/utilities/functions.php"
-        ]
-    },
-    "extra": {
-        "laravel": {
-            "dont-discover": [
-                "laravel/dusk",
-                "barryvdh/laravel-debugbar",
-                "rap2hpoutre/laravel-log-viewer"
-            ]
         }
     },
     "scripts": {
+        "post-autoload-dump": [
+            "Illuminate\\Foundation\\ComposerScripts::postAutoloadDump",
+            "@php artisan package:discover --ansi"
+        ],
+        "post-update-cmd": [
+            "@php artisan vendor:publish --tag=laravel-assets --ansi"
+        ],
         "post-root-package-install": [
             "@php -r \"file_exists('.env') || copy('.env.example', '.env');\""
         ],
         "post-create-project-cmd": [
-            "@php artisan key:generate"
-        ],
-        "post-install-cmd": [
-            "@php artisan passport:keys"
-        ],
-        "post-autoload-dump": [
-            "Illuminate\\Foundation\\ComposerScripts::postAutoloadDump",
-            "@php artisan package:discover"
+            "@php artisan key:generate --ansi"
         ]
     },
+    "extra": {
+        "laravel": {
+            "dont-discover": []
+        }
+    },
     "config": {
+        "optimize-autoloader": true,
         "preferred-install": "dist",
-        "sort-packages": true,
-        "optimize-autoloader": true
-    }
+        "sort-packages": true
+    },
+    "minimum-stability": "dev",
+    "prefer-stable": true
 }
