@@ -1,5 +1,30 @@
 <?php
 
+use Illuminate\Support\Facades\Auth;
+use Illuminate\Support\Facades\Route;
+use App\Http\Controllers\ExamController;
+use App\Http\Controllers\HomeController;
+use App\Http\Controllers\MarkController;
+use App\Http\Controllers\UserController;
+use App\Http\Controllers\EventController;
+use App\Http\Controllers\CourseController;
+use App\Http\Controllers\NoticeController;
+use App\Http\Controllers\RoutineController;
+use App\Http\Controllers\SectionController;
+use App\Http\Controllers\ExamRuleController;
+use App\Http\Controllers\SemesterController;
+use App\Http\Controllers\SyllabusController;
+use App\Http\Controllers\GradeRuleController;
+use App\Http\Controllers\PromotionController;
+use App\Http\Controllers\AssignmentController;
+use App\Http\Controllers\AttendanceController;
+use App\Http\Controllers\SchoolClassController;
+use App\Http\Controllers\GradingSystemController;
+use App\Http\Controllers\SchoolSessionController;
+use App\Http\Controllers\AcademicSettingController;
+use App\Http\Controllers\AssignedTeacherController;
+use App\Http\Controllers\Auth\UpdatePasswordController;
+
 /*
 |--------------------------------------------------------------------------
 | Web Routes
@@ -17,256 +42,138 @@ Route::get('/', function () {
 
 Auth::routes();
 
-Route::middleware(['auth', 'master'])->group(function () {
-    Route::get('/masters', 'MasterController@index')->name('masters.index');
-    Route::resource('/schools', 'SchoolController')->only(['index', 'edit', 'store', 'update']);
-});
-
-Route::get('/home', 'HomeController@index')->name('home');
-
 Route::middleware(['auth'])->group(function () {
-    Route::get('logs', '\Rap2hpoutre\LaravelLogViewer\LogViewerController@index');
-    // Route::get('/view-attendance/section/{section_id}',function($section_id){
-    //   if($section_id > 0){
-    //     $attendances = App\Attendance::with(['student'])->where('section_id', $section_id)->get();
-    //   }
-    // });
-    Route::get('attendances/students/{teacher_id}/{course_id}/{exam_id}/{section_id}', 'AttendanceController@addStudentsToCourseBeforeAtt')->middleware(['teacher']);
-    Route::get('attendances/{section_id}/{student_id}/{exam_id}', 'AttendanceController@index');
-    Route::get('attendances/{section_id}', 'AttendanceController@sectionIndex')->middleware(['teacher']);
-    Route::post('attendance/take-attendance', 'AttendanceController@store')->middleware(['teacher']);
-    Route::get('attendance/adjust/{student_id}', 'AttendanceController@adjust')->middleware(['teacher']);
-    Route::post('attendance/adjust', 'AttendanceController@adjustPost')->middleware(['teacher']);
-});
 
-Route::middleware(['auth', 'teacher'])->prefix('grades')->group(function () {
-    Route::get('all-exams-grade', 'GradeController@allExamsGrade');
-    Route::get('section/{section_id}', 'GradeController@gradesOfSection');
-    Route::get('t/{teacher_id}/{course_id}/{exam_id}/{section_id}', 'GradeController@tindex')->name('teacher-grade');
-    Route::get('c/{teacher_id}/{course_id}/{exam_id}/{section_id}', 'GradeController@cindex');
-    Route::post('calculate-marks', 'GradeController@calculateMarks');
-    Route::post('save-grade', 'GradeController@update');
-});
-
-Route::get('grades/{student_id}', 'GradeController@index')->middleware(['auth', 'teacher.student']);
-
-Route::middleware(['auth', 'accountant'])->prefix('fees')->name('fees.')->group(function () {
-    Route::get('all', 'FeeController@index');
-    Route::get('create', 'FeeController@create');
-    Route::post('create', 'FeeController@store');
-});
-
-Route::middleware(['auth', 'admin'])->group(function () {
-    Route::get('/settings', 'SettingController@index')->name('settings.index');
-    Route::get('gpa/create-gpa', 'GradesystemController@create');
-    Route::post('create-gpa', 'GradesystemController@store');
-    Route::post('gpa/delete', 'GradesystemController@destroy');
-});
-
-Route::middleware(['auth', 'teacher'])->group(function () {
-    Route::get('gpa/all-gpa', 'GradesystemController@index');
-});
-
-Route::middleware(['auth'])->group(function () {
-    if ('production' != config('app.env')) {
-        Route::get('user/config/impersonate', 'UserController@impersonateGet');
-        Route::post('user/config/impersonate', 'UserController@impersonate');
-    }
-
-    Route::get('users/{school_code}/{student_code}/{teacher_code}', 'UserController@index');
-    Route::get('users/{school_code}/{role}', 'UserController@indexOther');
-    Route::get('user/{user_code}', 'UserController@show');
-    Route::get('user/config/change_password', 'UserController@changePasswordGet');
-    Route::post('user/config/change_password', 'UserController@changePasswordPost');
-    Route::get('section/students/{section_id}', 'UserController@sectionStudents');
-
-    Route::get('courses/{teacher_id}/{section_id}', 'CourseController@index');
-});
-
-Route::middleware(['auth', 'teacher'])->group(function () {
-    Route::get('course/students/{teacher_id}/{course_id}/{exam_id}/{section_id}', 'CourseController@course');
-    Route::post('courses/create', 'CourseController@create');
-    // Route::post('courses/save-under-exam', 'CourseController@update');
-    Route::post('courses/store', 'CourseController@store');
-    Route::post('courses/save-configuration', 'CourseController@saveConfiguration');
-});
-
-Route::middleware(['auth', 'admin'])->prefix('academic')->name('academic.')->group(function () {
-    Route::get('syllabus', 'SyllabusController@index');
-    Route::get('syllabus/{class_id}', 'SyllabusController@create');
-    Route::get('notice', 'NoticeController@create');
-    Route::get('event', 'EventController@create');
-    Route::get('certificate', 'CertificateController@create');
-    Route::get('routine', 'RoutineController@index');
-    Route::get('routine/{section_id}', 'RoutineController@create');
-    Route::prefix('remove')->name('remove.')->group(function () {
-        Route::get('syllabus/{id}', 'SyllabusController@update');
-        Route::get('notice/{id}', 'NoticeController@update');
-        Route::get('event/{id}', 'EventController@update');
-        Route::get('certificate/{id}', 'CertificateController@update');
-        Route::get('routine/{id}', 'RoutineController@update');
-    });
-});
-
-Route::middleware(['auth', 'student'])->group(function () {
-    Route::get('user/{id}/notifications', 'NotificationController@index');
-    Route::get('academic/student/certificates', 'CertificateController@index');
-});
-
-Route::middleware(['auth', 'admin'])->prefix('exams')->name('exams.')->group(function () {
-    Route::get('/', 'ExamController@index');
-    Route::get('create', 'ExamController@create');
-    Route::post('create', 'ExamController@store');
-    Route::post('activate-exam', 'ExamController@update');
-});
-
-Route::middleware(['auth', 'teacher'])->group(function () {
-    Route::get('exams/active', 'ExamController@indexActive');
-    Route::get('school/sections', 'SectionController@index');
-});
-
-Route::middleware(['auth', 'librarian'])->namespace('Library')->group(function () {
-    Route::prefix('library')->name('library.')->group(function () {
-        Route::resource('books', 'BookController');
-    });
-});
-
-Route::middleware(['auth', 'librarian'])->prefix('library')->name('library.issued-books.')->group(function () {
-    Route::get('issue-books', 'IssuedbookController@create')->name('create');
-    Route::post('issue-books', 'IssuedbookController@store')->name('store');
-    Route::get('issued-books', 'IssuedbookController@index')->name('index');
-    Route::post('save_as_returned', 'IssuedbookController@update');
-});
-
-Route::middleware(['auth', 'accountant'])->prefix('accounts')->name('accounts.')->group(function () {
-    Route::get('sectors', 'AccountController@sectors')->name('sectors.index');
-    Route::post('create-sector', 'AccountController@storeSector')->name('sectors.create');
-    Route::get('edit-sector/{sector}', 'AccountController@editSector')->name('sectors.edit');
-    Route::patch('update-sector/{sector}', 'AccountController@updateSector')->name('sectors.update');
-    //Route::get('delete-sector/{sector}','AccountController@deleteSector')->name('sectors.delete');
-
-    Route::get('income', 'AccountController@income');
-    Route::post('create-income', 'AccountController@storeIncome');
-    Route::get('income-list', 'AccountController@listIncome');
-    Route::post('list-income', 'AccountController@postIncome');
-    Route::get('edit-income/{id}', 'AccountController@editIncome');
-    Route::post('update-income', 'AccountController@updateIncome');
-    Route::get('delete-income/{id}', 'AccountController@deleteIncome');
-
-    Route::get('expense', 'AccountController@expense');
-    Route::post('create-expense', 'AccountController@storeExpense');
-    Route::get('expense-list', 'AccountController@listExpense');
-    Route::post('list-expense', 'AccountController@postExpense');
-    Route::get('edit-expense/{id}', 'AccountController@editExpense');
-    Route::post('update-expense', 'AccountController@updateExpense');
-    Route::get('delete-expense/{id}', 'AccountController@deleteExpense');
-});
-
-Route::middleware(['auth', 'master'])->group(function () {
-    Route::get('register/admin/{id}/{code}', function ($id, $code) {
-        session([
-        'register_role' => 'admin',
-        'register_school_id' => $id,
-        'register_school_code' => $code,
-        ]);
-
-        return redirect()->route('register');
-    });
-    Route::post('register/admin', 'UserController@storeAdmin');
-    Route::get('master/activate-admin/{id}', 'UserController@activateAdmin');
-    Route::get('master/deactivate-admin/{id}', 'UserController@deactivateAdmin');
-    Route::get('school/admin-list/{school_id}', 'SchoolController@show');
-});
-
-Route::middleware(['auth', 'admin'])->group(function () {
     Route::prefix('school')->name('school.')->group(function () {
-        Route::post('add-class', 'MyclassController@store');
-        Route::post('add-section', 'SectionController@store');
-        Route::post('add-department', 'SchoolController@addDepartment');
-        Route::get('promote-students/{section_id}', 'UserController@promoteSectionStudents');
-        Route::post('promote-students', 'UserController@promoteSectionStudentsPost');
-        Route::post('theme', 'SchoolController@changeTheme');
-        Route::post('set-ignore-sessions', 'SchoolController@setIgnoreSessions');
+        Route::post('session/create', [SchoolSessionController::class, 'store'])->name('session.store');
+        Route::post('session/browse', [SchoolSessionController::class, 'browse'])->name('session.browse');
+
+        Route::post('semester/create', [SemesterController::class, 'store'])->name('semester.create');
+        Route::post('final-marks-submission-status/update', [AcademicSettingController::class, 'updateFinalMarksSubmissionStatus'])->name('final.marks.submission.status.update');
+
+        Route::post('attendance/type/update', [AcademicSettingController::class, 'updateAttendanceType'])->name('attendance.type.update');
+
+        // Class
+        Route::post('class/create', [SchoolClassController::class, 'store'])->name('class.create');
+        Route::post('class/update', [SchoolClassController::class, 'update'])->name('class.update');
+
+        // Sections
+        Route::post('section/create', [SectionController::class, 'store'])->name('section.create');
+        Route::post('section/update', [SectionController::class, 'update'])->name('section.update');
+
+        // Courses
+        Route::post('course/create', [CourseController::class, 'store'])->name('course.create');
+        Route::post('course/update', [CourseController::class, 'update'])->name('course.update');
+
+        // Teacher
+        Route::post('teacher/create', [UserController::class, 'storeTeacher'])->name('teacher.create');
+        Route::post('teacher/update', [UserController::class, 'updateTeacher'])->name('teacher.update');
+        Route::post('teacher/assign', [AssignedTeacherController::class, 'store'])->name('teacher.assign');
+
+        // Student
+        Route::post('student/create', [UserController::class, 'storeStudent'])->name('student.create');
+        Route::post('student/update', [UserController::class, 'updateStudent'])->name('student.update');
     });
 
-    Route::prefix('register')->name('register.')->group(function () {
-        Route::get('student', 'UserController@redirectToRegisterStudent');
-        Route::get('teacher', function () {
-            $departments = \App\Department::where('school_id', \Auth::user()->school_id)->get();
-            $classes = \App\Myclass::where('school_id', \Auth::user()->school->id)->pluck('id');
-            $sections = \App\Section::with('class')->whereIn('class_id', $classes)->get();
-            session([
-        'register_role' => 'teacher',
-        'departments' => $departments,
-        'register_sections' => $sections,
-      ]);
 
-            return redirect()->route('register');
-        });
-        Route::get('accountant', function () {
-            session(['register_role' => 'accountant']);
+    Route::get('/home', [HomeController::class, 'index'])->name('home');
 
-            return redirect()->route('register');
-        });
-        Route::get('librarian', function () {
-            session(['register_role' => 'librarian']);
+    // Attendance
+    Route::get('/attendances', [AttendanceController::class, 'index'])->name('attendance.index');
+    Route::get('/attendances/view', [AttendanceController::class, 'show'])->name('attendance.list.show');
+    Route::get('/attendances/take', [AttendanceController::class, 'create'])->name('attendance.create.show');
+    Route::post('/attendances', [AttendanceController::class, 'store'])->name('attendances.store');
 
-            return redirect()->route('register');
-        });
-        Route::post('student', 'UserController@store');
-        Route::post('teacher', 'UserController@storeTeacher');
-        Route::post('accountant', 'UserController@storeAccountant');
-        Route::post('librarian', 'UserController@storeLibrarian');
-    });
-    Route::get('edit/course/{id}', 'CourseController@edit');
-    Route::post('edit/course/{id}', 'CourseController@updateNameAndTime');
-});
+    // Classes and sections
+    Route::get('/classes', [SchoolClassController::class, 'index']);
+    Route::get('/class/edit/{id}', [SchoolClassController::class, 'edit'])->name('class.edit');
+    Route::get('/sections', [SectionController::class, 'getByClassId'])->name('get.sections.courses.by.classId');
+    Route::get('/section/edit/{id}', [SectionController::class, 'edit'])->name('section.edit');
 
-//use PDF;
-Route::middleware(['auth', 'master.admin'])->group(function () {
-    Route::get('edit/user/{id}', 'UserController@edit');
-    Route::post('edit/user', 'UserController@update');
-    Route::post('upload/file', 'UploadController@upload');
-    Route::post('users/import/user-xlsx', 'UploadController@import');
-    Route::get('users/export/students-xlsx', 'UploadController@export');
-    //   Route::get('pdf/profile/{user_id}',function($user_id){
-//     $data = App\User::find($user_id);
-//     PDF::setOptions(['defaultFont' => 'sans-serif', 'isHtml5ParserEnabled' => true]);
-//     $pdf = PDF::loadView('pdf.profile-pdf', ['user' => $data]);
-// 		return $pdf->stream('profile.pdf');
-//   });
-//   Route::get('pdf/result/{user_id}/{exam_id}',function($user_id, $exam_id){
-//     $data = App\User::find($user_id);
-//     $grades = App\Grade::with('exam')->where('student_id', $user_id)->where('exam_id',$exam_id)->latest()->get();
-//     PDF::setOptions(['defaultFont' => 'sans-serif', 'isHtml5ParserEnabled' => true]);
-//     $pdf = PDF::loadView('pdf.result-pdf', ['grades' => $grades, 'user'=>$data]);
-// 		return $pdf->stream('result.pdf');
-//   });
-});
-Route::middleware(['auth', 'teacher'])->group(function () {
-    Route::post('calculate-marks', 'GradeController@calculateMarks');
-    Route::post('message/students', 'NotificationController@store');
-});
-// Route::middleware(['auth'])->group(function (){
-//   Route::get('download/pdf', function(){
-//     $pathToFile = public_path('storage/Bano-EducationandAspiration.pdf');
-//     return response()->download($pathToFile);
-//   });
-// });
+    // Teachers
+    Route::get('/teachers/add', function () {
+        return view('teachers.add');
+    })->name('teacher.create.show');
+    Route::get('/teachers/edit/{id}', [UserController::class, 'editTeacher'])->name('teacher.edit.show');
+    Route::get('/teachers/view/list', [UserController::class, 'getTeacherList'])->name('teacher.list.show');
+    Route::get('/teachers/view/profile/{id}', [UserController::class, 'showTeacherProfile'])->name('teacher.profile.show');
 
-// View Emails - in browser
-Route::prefix('emails')->group(function () {
-    // Welcome Email
-    Route::get('/welcome', function () {
-        $user = App\User::find(1);
-        $password = 'ABCXYZ';
+    //Students
+    Route::get('/students/add', [UserController::class, 'createStudent'])->name('student.create.show');
+    Route::get('/students/edit/{id}', [UserController::class, 'editStudent'])->name('student.edit.show');
+    Route::get('/students/view/list', [UserController::class, 'getStudentList'])->name('student.list.show');
+    Route::get('/students/view/profile/{id}', [UserController::class, 'showStudentProfile'])->name('student.profile.show');
+    Route::get('/students/view/attendance/{id}', [AttendanceController::class, 'showStudentAttendance'])->name('student.attendance.show');
 
-        return new App\Mail\SendWelcomeEmailToUser($user, $password);
-    });
-});
+    // Marks
+    Route::get('/marks/create', [MarkController::class, 'create'])->name('course.mark.create');
+    Route::post('/marks/store', [MarkController::class, 'store'])->name('course.mark.store');
+    Route::get('/marks/results', [MarkController::class, 'index'])->name('course.mark.list.show');
+    // Route::get('/marks/view', function () {
+    //     return view('marks.view');
+    // });
+    Route::get('/marks/view', [MarkController::class, 'showCourseMark'])->name('course.mark.show');
+    Route::get('/marks/final/submit', [MarkController::class, 'showFinalMark'])->name('course.final.mark.submit.show');
+    Route::post('/marks/final/submit', [MarkController::class, 'storeFinalMark'])->name('course.final.mark.submit.store');
 
-Route::middleware(['auth', 'student'])->prefix('stripe')->group(function () {
-    Route::get('charge', 'CashierController@index');
-    Route::post('charge', 'CashierController@store');
-    Route::get('receipts', 'PaymentController@index');
+    // Exams
+    Route::get('/exams/view', [ExamController::class, 'index'])->name('exam.list.show');
+    // Route::get('/exams/view/history', function () {
+    //     return view('exams.history');
+    // });
+    Route::post('/exams/create', [ExamController::class, 'store'])->name('exam.create');
+    // Route::post('/exams/delete', [ExamController::class, 'delete'])->name('exam.delete');
+    Route::get('/exams/create', [ExamController::class, 'create'])->name('exam.create.show');
+    Route::get('/exams/add-rule', [ExamRuleController::class, 'create'])->name('exam.rule.create');
+    Route::post('/exams/add-rule', [ExamRuleController::class, 'store'])->name('exam.rule.store');
+    Route::get('/exams/edit-rule', [ExamRuleController::class, 'edit'])->name('exam.rule.edit');
+    Route::post('/exams/edit-rule', [ExamRuleController::class, 'update'])->name('exam.rule.update');
+    Route::get('/exams/view-rule', [ExamRuleController::class, 'index'])->name('exam.rule.show');
+    Route::get('/exams/grade/create', [GradingSystemController::class, 'create'])->name('exam.grade.system.create');
+    Route::post('/exams/grade/create', [GradingSystemController::class, 'store'])->name('exam.grade.system.store');
+    Route::get('/exams/grade/view', [GradingSystemController::class, 'index'])->name('exam.grade.system.index');
+    Route::get('/exams/grade/add-rule', [GradeRuleController::class, 'create'])->name('exam.grade.system.rule.create');
+    Route::post('/exams/grade/add-rule', [GradeRuleController::class, 'store'])->name('exam.grade.system.rule.store');
+    Route::get('/exams/grade/view-rules', [GradeRuleController::class, 'index'])->name('exam.grade.system.rule.show');
+    Route::post('/exams/grade/delete-rule', [GradeRuleController::class, 'destroy'])->name('exam.grade.system.rule.delete');
+
+    // Promotions
+    Route::get('/promotions/index', [PromotionController::class, 'index'])->name('promotions.index');
+    Route::get('/promotions/promote', [PromotionController::class, 'create'])->name('promotions.create');
+    Route::post('/promotions/promote', [PromotionController::class, 'store'])->name('promotions.store');
+
+    // Academic settings
+    Route::get('/academics/settings', [AcademicSettingController::class, 'index']);
+
+    // Calendar events
+    Route::get('calendar-event', [EventController::class, 'index'])->name('events.show');
+    Route::post('calendar-crud-ajax', [EventController::class, 'calendarEvents'])->name('events.crud');
+
+    // Routines
+    Route::get('/routine/create', [RoutineController::class, 'create'])->name('section.routine.create');
+    Route::get('/routine/view', [RoutineController::class, 'show'])->name('section.routine.show');
+    Route::post('/routine/store', [RoutineController::class, 'store'])->name('section.routine.store');
+
+    // Syllabus
+    Route::get('/syllabus/create', [SyllabusController::class, 'create'])->name('class.syllabus.create');
+    Route::post('/syllabus/create', [SyllabusController::class, 'store'])->name('syllabus.store');
+    Route::get('/syllabus/index', [SyllabusController::class, 'index'])->name('course.syllabus.index');
+
+    // Notices
+    Route::get('/notice/create', [NoticeController::class, 'create'])->name('notice.create');
+    Route::post('/notice/create', [NoticeController::class, 'store'])->name('notice.store');
+
+    // Courses
+    Route::get('courses/teacher/index', [AssignedTeacherController::class, 'getTeacherCourses'])->name('course.teacher.list.show');
+    Route::get('courses/student/index/{student_id}', [CourseController::class, 'getStudentCourses'])->name('course.student.list.show');
+    Route::get('course/edit/{id}', [CourseController::class, 'edit'])->name('course.edit');
+
+    // Assignment
+    Route::get('courses/assignments/index', [AssignmentController::class, 'getCourseAssignments'])->name('assignment.list.show');
+    Route::get('courses/assignments/create', [AssignmentController::class, 'create'])->name('assignment.create');
+    Route::post('courses/assignments/create', [AssignmentController::class, 'store'])->name('assignment.store');
+
+    // Update password
+    Route::get('password/edit', [UpdatePasswordController::class, 'edit'])->name('password.edit');
+    Route::post('password/edit', [UpdatePasswordController::class, 'update'])->name('password.update');
 });
