@@ -13,7 +13,7 @@ class Kernel extends ConsoleKernel
      * @var array
      */
     protected $commands = [
-        // \jdavidbakr\LaravelCacheGarbageCollector\LaravelCacheGarbageCollector::class,
+        //
     ];
 
     /**
@@ -24,8 +24,7 @@ class Kernel extends ConsoleKernel
      */
     protected function schedule(Schedule $schedule)
     {
-        //$schedule->command('cache:gc')
-        //          ->everyMinute();
+        // $schedule->command('inspire')->hourly();
     }
 
     /**
