@@ -2,11 +2,9 @@
 
 namespace App\Http\Controllers\Auth;
 
-use Illuminate\Support\Arr;
-use Illuminate\Support\Facades\Auth;
 use App\Http\Controllers\Controller;
+use App\Providers\RouteServiceProvider;
 use Illuminate\Foundation\Auth\AuthenticatesUsers;
-use Illuminate\Http\Request;
 
 class LoginController extends Controller
 {
@@ -24,6 +22,13 @@ class LoginController extends Controller
     use AuthenticatesUsers;
 
     /**
+     * Where to redirect users after login.
+     *
+     * @var string
+     */
+    protected $redirectTo = RouteServiceProvider::HOME;
+
+    /**
      * Create a new controller instance.
      *
      * @return void
@@ -31,27 +36,5 @@ class LoginController extends Controller
     public function __construct()
     {
         $this->middleware('guest')->except('logout');
-    }
-
-    public function redirectTo()
-    {
-        $role = Auth::user()->role;
-
-        return ($role == 'master')?'/masters':'/home';
-    }
-
-    public function username()
-    {
-        $login = request()->input('email');
-        $field = filter_var($login, FILTER_VALIDATE_EMAIL) ? 'email' : 'phone_number';
-        request()->merge([$field => $login]);
-        return $field;
-    }
-
-    public function credentials(Request $request)
-    {
-        $credentials = $request->only($this->username(), 'password');
-        $credentials = Arr::add($credentials, 'active', '1');
-        return $credentials;
     }
 }
