@@ -2,45 +2,40 @@
 
 namespace App\Http\Controllers;
 
-use App\Course;
-use App\Http\Resources\CourseResource;
+use App\Http\Controllers\Controller;
+use App\Models\Course;
 use Illuminate\Http\Request;
-use App\Http\Requests\Course\SaveConfigurationRequest;
-use App\Http\Traits\GradeTrait;
-use App\Services\Course\CourseService;
+use App\Traits\SchoolSession;
+use App\Interfaces\CourseInterface;
+use App\Http\Requests\CourseStoreRequest;
+use App\Interfaces\SchoolSessionInterface;
+use App\Repositories\PromotionRepository;
 
 class CourseController extends Controller
 {
-    use GradeTrait;
-    protected $courseService;
+    use SchoolSession;
+    protected $schoolCourseRepository;
+    protected $schoolSessionRepository;
 
-    public function __construct(CourseService $courseService){
-      $this->courseService = $courseService;
+    /**
+    * Create a new Controller instance
+    * 
+    * @param CourseInterface $schoolCourseRepository
+    * @return void
+    */
+    public function __construct(SchoolSessionInterface $schoolSessionRepository, CourseInterface $schoolCourseRepository) {
+        $this->schoolSessionRepository = $schoolSessionRepository;
+        $this->schoolCourseRepository = $schoolCourseRepository;
     }
+
     /**
      * Display a listing of the resource.
      *
      * @return \Illuminate\Http\Response
      */
-    public function index($teacher_id, $section_id){
-      if($this->courseService->isCourseOfTeacher($teacher_id)) {
-        $courses = $this->courseService->getCoursesByTeacher($teacher_id);
-        $exams = $this->courseService->getExamsBySchoolId();
-        $view = 'course.teacher-course';
-
-      } else if($this->courseService->isCourseOfStudentOfASection($section_id)) {
-        $courses = $this->courseService->getCoursesBySection($section_id);
-        $view = 'course.class-course';
-        $exams = [];
-
-      } else if($this->courseService->isCourseOfASection($section_id)) {
-        $courses = $this->courseService->getCoursesBySection($section_id);
-        $exams = $this->courseService->getExamsBySchoolId();
-        $view = 'course.class-course';
-      } else {
-        return redirect('home');
-      }
-      return view($view,compact('courses','exams'));
+    public function index()
+    {
+        //
     }
 
     /**
@@ -54,99 +49,86 @@ class CourseController extends Controller
     }
 
     /**
-     * Show the form for creating a new resource.
-     *
-     * @return \Illuminate\Http\Response
-     */
-    public function course($teacher_id,$course_id,$exam_id,$section_id)
-    {
-      $this->addStudentsToCourse($teacher_id,$course_id,$exam_id,$section_id);
-      $students = $this->courseService->getStudentsFromGradeByCourseAndExam($course_id, $exam_id);
-
-      return view('course.students', compact('students','teacher_id','section_id'));
-    }
-
-    /**
      * Store a newly created resource in storage.
      *
-     * @param  \Illuminate\Http\Request  $request
+     * @param  CourseStoreRequest $request
      * @return \Illuminate\Http\Response
      */
-    public function store(Request $request)
+    public function store(CourseStoreRequest $request)
     {
-      try{
-        $this->courseService->addCourse($request);
-      } catch (\Exception $ex){
-        return __('Could not add course.');
-      }
-      return back()->with('status', __('Created'));
-    }
+        try {
+            $this->schoolCourseRepository->create($request->validated());
 
-    /**
-     * @param SaveConfigurationRequest $request
-     * @return \Illuminate\Http\RedirectResponse
-     */
-    public function saveConfiguration(SaveConfigurationRequest $request){
-      try{
-        $this->courseService->saveConfiguration($request);
-      } catch (\Exception $ex){
-        return __('Could not save configuration.');
-      }
-      return back()->with('status', __('Saved'));
+            return back()->with('status', 'Course creation was successful!');
+        } catch (\Exception $e) {
+            return back()->withError($e->getMessage());
+        }
     }
 
     /**
      * Display the specified resource.
      *
-     * @param  int  $id
      * @return \Illuminate\Http\Response
      */
-    public function show($id)
-    {
-      return new CourseResource(Course::find($id));
+    public function getStudentCourses($student_id) {
+        $current_school_session_id = $this->getSchoolCurrentSession();
+        $promotionRepository = new PromotionRepository();
+        $class_info = $promotionRepository->getPromotionInfoById($current_school_session_id, $student_id);
+        $courses = $this->schoolCourseRepository->getByClassId($class_info->class_id);
+
+        $data = [
+            'class_info'    => $class_info,
+            'courses'       => $courses,
+        ];
+        return view('courses.student', $data);
     }
 
     /**
      * Show the form for editing the specified resource.
      *
-     * @param  int  $id
+     * @param  $course_id
      * @return \Illuminate\Http\Response
      */
-    public function edit($id)
+    public function edit($course_id)
     {
-      $course = Course::find($id);
-      return view('course.edit', ['course'=>$course]);
+        $current_school_session_id = $this->getSchoolCurrentSession();
+
+        $course = $this->schoolCourseRepository->findById($course_id);
+
+        $data = [
+            'current_school_session_id' => $current_school_session_id,
+            'course'                    => $course,
+            'course_id'                 => $course_id,
+        ];
+
+        return view('courses.edit', $data);
     }
 
     /**
      * Update the specified resource in storage.
      *
      * @param  \Illuminate\Http\Request  $request
-     * @param  int  $id
      * @return \Illuminate\Http\Response
      */
-    public function updateNameAndTime(Request $request, $id)
+    public function update(Request $request)
     {
-      $request->validate([
-        'course_name' => 'required|string',
-        'course_time' => 'required|string',
-      ]);
-      $this->courseService->updateCourseInfo($id, $request);
-      return back()->with('status', __('Saved'));
+        try {
+            $this->schoolCourseRepository->update($request);
+
+            return back()->with('status', 'Course update was successful!');
+        } catch (\Exception $e) {
+            return back()->withError($e->getMessage());
+        }
     }
 
     /**
      * Remove the specified resource from storage.
      *
-     * @param  int  $id
+     * @param  \App\Models\Course  $course
      * @return \Illuminate\Http\Response
      */
-    public function destroy($id)
+    public function destroy(Course $course)
     {
-      return (Course::destroy($id))?response()->json([
-        'status' => 'success'
-      ]):response()->json([
-        'status' => 'error'
-      ]);
+        //
     }
 }
