@@ -2,34 +2,67 @@
 
 namespace App\Http\Controllers;
 
+use App\Http\Controllers\Controller;
+use App\Http\Requests\ExamStoreRequest;
+use App\Models\Exam;
 use Illuminate\Http\Request;
-use App\Services\Exam\ExamService;
-use App\Http\Requests\Exam\CreateExamRequest;
+use App\Traits\SchoolSession;
+use App\Interfaces\SemesterInterface;
+use App\Interfaces\SchoolClassInterface;
+use App\Interfaces\SchoolSessionInterface;
+use App\Repositories\AssignedTeacherRepository;
+use App\Repositories\ExamRepository;
 
 class ExamController extends Controller
 {
-    protected $examService;
+    use SchoolSession;
 
-    public function __construct(ExamService $examService){
-        $this->examService = $examService;
+    protected $schoolClassRepository;
+    protected $semesterRepository;
+    protected $schoolSessionRepository;
+
+    public function __construct(SchoolSessionInterface $schoolSessionRepository, SchoolClassInterface $schoolClassRepository, SemesterInterface $semesterRepository)
+    {
+        $this->schoolSessionRepository = $schoolSessionRepository;
+        $this->schoolClassRepository = $schoolClassRepository;
+        $this->semesterRepository = $semesterRepository;
     }
     /**
      * Display a listing of the resource.
      *
+     * @param  \Illuminate\Http\Request  $request
      * @return \Illuminate\Http\Response
      */
-    public function index()
+    public function index(Request $request)
     {
-        $exams = $this->examService->getLatestExamsBySchoolIdWithPagination();
-        return view('exams.all',compact('exams'));
-    }
+        $class_id = $request->query('class_id', 0);
+        $semester_id = $request->query('semester_id', 0);
 
-    public function indexActive(){
-        $exams = $this->examService->getActiveExamsBySchoolId();
-        $this->examService->examIds = $exams->pluck('id')->toArray();
-        $courses = $this->examService->getCoursesByExamIds();
+        $current_school_session_id = $this->getSchoolCurrentSession();
 
-        return view('exams.active',compact('exams','courses'));
+        $semesters = $this->semesterRepository->getAll($current_school_session_id);
+
+        $school_classes = $this->schoolClassRepository->getAllBySession($current_school_session_id);
+
+        $examRepository = new ExamRepository();
+
+        $exams = $examRepository->getAll($current_school_session_id, $semester_id, $class_id);
+
+        $assignedTeacherRepository = new AssignedTeacherRepository();
+
+        $teacher_id = (auth()->user()->role == "teacher")?auth()->user()->id : 0;
+
+        $teacherCourses = $assignedTeacherRepository->getTeacherCourses($current_school_session_id, $teacher_id, $semester_id);
+
+        $data = [
+            'current_school_session_id' => $current_school_session_id,
+            'semesters'                 => $semesters,
+            'classes'                   => $school_classes,
+            'exams'                     => $exams,
+            'teacher_courses'           => $teacherCourses,
+        ];
+
+        return view('exams.index', $data);
     }
 
     /**
@@ -39,38 +72,59 @@ class ExamController extends Controller
      */
     public function create()
     {
-        $classes = $this->examService->getClassesBySchoolId();
-        $already_assigned_classes = $this->examService->getAlreadyAssignedClasses();
-        return view('exams.add',compact('classes','already_assigned_classes'));
+        $current_school_session_id = $this->getSchoolCurrentSession();
+
+        $semesters = $this->semesterRepository->getAll($current_school_session_id);
+
+        if(auth()->user()->role == "teacher") {
+            $teacher_id = auth()->user()->id;
+            $assigned_classes = $this->schoolClassRepository->getAllBySessionAndTeacher($current_school_session_id, $teacher_id);
+
+            $school_classes = [];
+            $i = 0;
+
+            foreach($assigned_classes as $assigned_class) {
+                $school_classes[$i] = $assigned_class->schoolClass;
+                $i++;
+            }
+        } else {
+            $school_classes = $this->schoolClassRepository->getAllBySession($current_school_session_id);
+        }
+
+        $data = [
+            'current_school_session_id' => $current_school_session_id,
+            'semesters'                 => $semesters,
+            'classes'                   => $school_classes,
+        ];
+
+        return view('exams.create', $data);
     }
 
     /**
      * Store a newly created resource in storage.
      *
-     * @param  \Illuminate\Http\Request  $request
+     * @param  ExamStoreRequest  $request
      * @return \Illuminate\Http\Response
      */
-    public function store(CreateExamRequest $request)
+    public function store(ExamStoreRequest $request)
     {
-        
-        $this->examService->request = $request;
-        try{
-            $this->examService->storeExam();
-        } catch (\Exception $e){
-            return 'Error: '. $e->getMessage();
+        try {
+            $examRepository = new ExamRepository();
+            $examRepository->create($request->validated());
+
+            return back()->with('status', 'Exam creation was successful!');
+        } catch (\Exception $e) {
+            return back()->withError($e->getMessage());
         }
-        
-        //return $this->cindex($course_id, $exam_id, $teacher_id);
-        return back()->with('status', __('Created'));
     }
 
     /**
      * Display the specified resource.
      *
-     * @param  int  $id
+     * @param  \App\Models\Exam  $exam
      * @return \Illuminate\Http\Response
      */
-    public function show($id)
+    public function show(Exam $exam)
     {
         //
     }
@@ -78,10 +132,10 @@ class ExamController extends Controller
     /**
      * Show the form for editing the specified resource.
      *
-     * @param  int  $id
+     * @param  \App\Models\Exam  $exam
      * @return \Illuminate\Http\Response
      */
-    public function edit($id)
+    public function edit(Exam $exam)
     {
         //
     }
@@ -90,51 +144,29 @@ class ExamController extends Controller
      * Update the specified resource in storage.
      *
      * @param  \Illuminate\Http\Request  $request
-     * @param  int  $id
+     * @param  \App\Models\Exam  $exam
      * @return \Illuminate\Http\Response
      */
-    public function update(Request $request)
+    public function update(Request $request, Exam $exam)
     {
-        $request->validate([
-            'exam_id' => 'required|numeric',
-        ]);
-        try{
-            $this->examService->request = $request;
-            $this->examService->updateExam();
-        } catch (\Exception $e){
-            return 'Error: '. $e->getMessage();
-        }
-        return back()->with('status', __('Saved'));
-    }
-
-    /**
-     * Update the specified resource in storage.
-     *
-     * @param  \Illuminate\Http\Request  $request
-     * @param  int  $id
-     * @return \Illuminate\Http\Response
-     */
-    private function assignCoursesToExam()
-    {
-    //   $request->validate([
-    //     'course_id' => 'required|numeric',
-    //     'exam_id' => 'required|numeric',
-    //   ]);
-        
-        // $tb = Course::find($request->course_id);
-        // $tb->exam_id = $request->exam_id;
-        // $tb->save();
-        // return back()->with('status', 'Saved');
+        //
     }
 
     /**
      * Remove the specified resource from storage.
      *
-     * @param  int  $id
+     * @param  \App\Models\Exam  $exam
      * @return \Illuminate\Http\Response
      */
-    public function destroy($id)
+    public function destroy(Request $request)
     {
-        //
+        try {
+            $examRepository = new ExamRepository();
+            $examRepository->delete($request->exam_id);
+
+            return back()->with('status', 'Exam deletion was successful!');
+        } catch (\Exception $e) {
+            return back()->withError($e->getMessage());
+        }
     }
 }
