@@ -2,101 +2,64 @@
 
 namespace App\Http\Controllers;
 
-use Illuminate\Http\Request;
+use App\Http\Controllers\Controller;
+use App\Traits\SchoolSession;
+use App\Interfaces\UserInterface;
+use App\Repositories\NoticeRepository;
+use App\Interfaces\SchoolClassInterface;
+use App\Interfaces\SchoolSessionInterface;
+use App\Repositories\PromotionRepository;
 
 class HomeController extends Controller
 {
+    use SchoolSession;
+    protected $schoolSessionRepository;
+    protected $schoolClassRepository;
+    protected $userRepository;
     /**
      * Create a new controller instance.
      *
      * @return void
      */
-    public function __construct()
+    public function __construct(
+        UserInterface $userRepository, SchoolSessionInterface $schoolSessionRepository, SchoolClassInterface $schoolClassRepository)
     {
-        $this->middleware('auth');
+        // $this->middleware('auth');
+        $this->userRepository = $userRepository;
+        $this->schoolSessionRepository = $schoolSessionRepository;
+        $this->schoolClassRepository = $schoolClassRepository;
     }
 
     /**
      * Show the application dashboard.
      *
-     * @return \Illuminate\Http\Response
+     * @return \Illuminate\Contracts\Support\Renderable
      */
     public function index()
     {
-        
-        if (\Auth::user()->role != 'master') {
-            $minutes = 1440;// 24 hours = 1440 minutes
-            $school_id = \Auth::user()->school->id;
-            $classes = \Cache::remember('classes-'.$school_id, $minutes, function () use($school_id) {
-              return \App\Myclass::bySchool($school_id)
-                                ->pluck('id')
-                                ->toArray();
-            });
-            $totalStudents = \Cache::remember('totalStudents-'.$school_id, $minutes, function () use($school_id) {
-              return \App\User::bySchool($school_id)
-                              ->where('role','student')
-                              ->where('active', 1)
-                              ->count();
-            });
-            $totalTeachers = \Cache::remember('totalTeachers-'.$school_id, $minutes, function () use($school_id) {
-              return \App\User::bySchool($school_id)
-                              ->where('role','teacher')
-                              ->where('active', 1)
-                              ->count();
-            });
-            $totalBooks = \Cache::remember('totalBooks-'.$school_id, $minutes, function () use($school_id) {
-              return \App\Book::bySchool($school_id)->count();
-            });
-            $totalClasses = \Cache::remember('totalClasses-'.$school_id, $minutes, function () use($school_id) {
-              return \App\Myclass::bySchool($school_id)->count();
-            });
-            $totalSections = \Cache::remember('totalSections-'.$school_id, $minutes, function () use ($classes) {
-              return \App\Section::whereIn('class_id', $classes)->count();
-            });
-            $notices = \Cache::remember('notices-'.$school_id, $minutes, function () use($school_id) {
-              return \App\Notice::bySchool($school_id)
-                                ->where('active',1)
-                                ->get();
-            });
-            $events = \Cache::remember('events-'.$school_id, $minutes, function () use($school_id) {
-              return \App\Event::bySchool($school_id)
-                              ->where('active',1)
-                              ->get();
-            });
-            $routines = \Cache::remember('routines-'.$school_id, $minutes, function () use($school_id) {
-              return \App\Routine::bySchool($school_id)
-                                ->where('active',1)
-                                ->get();
-            });
-            $syllabuses = \Cache::remember('syllabuses-'.$school_id, $minutes, function () use($school_id) {
-              return \App\Syllabus::bySchool($school_id)
-                                  ->where('active',1)
-                                  ->get();
-            });
-            $exams = \Cache::remember('exams-'.$school_id, $minutes, function () use($school_id) {
-              return \App\Exam::bySchool($school_id)
-                              ->where('active',1)
-                              ->get();
-            });
-            // if(\Auth::user()->role == 'student')
-            //   $messageCount = \App\Notification::where('student_id',\Auth::user()->id)->count();
-            // else
-            //   $messageCount = 0;
-            return view('home',[
-              'totalStudents'=>$totalStudents,
-              'totalTeachers'=>$totalTeachers,
-              'totalBooks'=>$totalBooks,
-              'totalClasses'=>$totalClasses,
-              'totalSections'=>$totalSections,
-              'notices'=>$notices,
-              'events'=>$events,
-              'routines'=>$routines,
-              'syllabuses'=>$syllabuses,
-              'exams'=>$exams,
-              //'messageCount'=>$messageCount,
-            ]);
-        } else {
-            return redirect('/masters');
-        }
+        $current_school_session_id = $this->getSchoolCurrentSession();
+
+        $classCount = $this->schoolClassRepository->getAllBySession($current_school_session_id)->count();
+
+        $studentCount = $this->userRepository->getAllStudentsBySessionCount($current_school_session_id);
+
+        $promotionRepository = new PromotionRepository();
+
+        $maleStudentsBySession = $promotionRepository->getMaleStudentsBySessionCount($current_school_session_id);
+
+        $teacherCount = $this->userRepository->getAllTeachers()->count();
+
+        $noticeRepository = new NoticeRepository();
+        $notices = $noticeRepository->getAll($current_school_session_id);
+
+        $data = [
+            'classCount'    => $classCount,
+            'studentCount'  => $studentCount,
+            'teacherCount'  => $teacherCount,
+            'notices'       => $notices,
+            'maleStudentsBySession' => $maleStudentsBySession,
+        ];
+
+        return view('home', $data);
     }
 }
