@@ -2,428 +2,185 @@
 
 namespace App\Http\Controllers;
 
-use App\Department;
-use App\Myclass;
-use App\Section;
-use App\StudentInfo;
-use App\User;
+use App\Http\Controllers\Controller;
 use Illuminate\Http\Request;
-use Illuminate\Support\Facades\Auth;
-use Illuminate\Support\Facades\DB;
-use App\Http\Resources\UserResource;
-use Illuminate\Support\Facades\Hash;
-use App\Http\Requests\User\CreateUserRequest;
-use App\Http\Requests\User\UpdateUserRequest;
-use App\Http\Requests\User\CreateAdminRequest;
-use App\Http\Requests\User\CreateTeacherRequest;
-use App\Http\Requests\User\ChangePasswordRequest;
-use App\Http\Requests\User\ImpersonateUserRequest;
-use App\Http\Requests\User\CreateLibrarianRequest;
-use App\Http\Requests\User\CreateAccountantRequest;
-use App\Events\UserRegistered;
-use App\Events\StudentInfoUpdateRequested;
-use Illuminate\Support\Facades\Log;
-use App\Services\User\UserService;
-/**
- * Class UserController
- * @package App\Http\Controllers
- */
+use App\Traits\SchoolSession;
+use App\Interfaces\UserInterface;
+use App\Interfaces\SectionInterface;
+use App\Interfaces\SchoolClassInterface;
+use App\Repositories\PromotionRepository;
+use App\Http\Requests\StudentStoreRequest;
+use App\Http\Requests\TeacherStoreRequest;
+use App\Interfaces\SchoolSessionInterface;
+use App\Repositories\StudentParentInfoRepository;
+
 class UserController extends Controller
 {
-    protected $userService;
-    protected $user;
+    use SchoolSession;
+    protected $userRepository;
+    protected $schoolSessionRepository;
+    protected $schoolClassRepository;
+    protected $schoolSectionRepository;
 
-    public function __construct(UserService $userService, User $user){
-        $this->userService = $userService;
-        $this->user = $user;
+    public function __construct(UserInterface $userRepository, SchoolSessionInterface $schoolSessionRepository,
+    SchoolClassInterface $schoolClassRepository,
+    SectionInterface $schoolSectionRepository)
+    {
+        $this->middleware(['can:view users']);
+
+        $this->userRepository = $userRepository;
+        $this->schoolSessionRepository = $schoolSessionRepository;
+        $this->schoolClassRepository = $schoolClassRepository;
+        $this->schoolSectionRepository = $schoolSectionRepository;
     }
+    
     /**
-     * Display a listing of the resource.
+     * Store a newly created resource in storage.
      *
-     * @param $school_code
-     * @param $student_code
-     * @param $teacher_code
+     * @param  TeacherStoreRequest $request
      * @return \Illuminate\Http\Response
      */
-    public function index($school_code, $student_code, $teacher_code){
-        session()->forget('section-attendance');
-        
-        if($this->userService->isListOfStudents($school_code, $student_code))
-            return $this->userService->indexView('list.student-list', $this->userService->getStudents());
-        else if($this->userService->isListOfTeachers($school_code, $teacher_code))
-            return $this->userService->indexView('list.teacher-list',$this->userService->getTeachers());
-        else
-            return view('home');
-    }
-
-    /**
-     * @param $school_code
-     * @param $role
-     * @return \Illuminate\Contracts\View\Factory|\Illuminate\View\View
-     */
-    public function indexOther($school_code, $role){
-        if($this->userService->isAccountant($role))
-            return $this->userService->indexOtherView('accounts.accountant-list', $this->userService->getAccountants());
-        else if($this->userService->isLibrarian($role))
-            return $this->userService->indexOtherView('library.librarian-list', $this->userService->getLibrarians());
-        else
-            return view('home');
-    }
-
-    /**
-     * @return \Illuminate\Http\RedirectResponse
-     */
-    public function redirectToRegisterStudent()
+    public function storeTeacher(TeacherStoreRequest $request)
     {
-        $classes = Myclass::query()
-            ->bySchool(\Auth::user()->school->id)
-            ->pluck('id');
+        try {
+            $this->userRepository->createTeacher($request->validated());
 
-        $sections = Section::with('class')
-            ->whereIn('class_id', $classes)
-            ->get();
-
-        session([
-            'register_role' => 'student',
-            'register_sections' => $sections,
-        ]);
-
-        return redirect()->route('register');
-    }
-
-    /**
-     * @param $section_id
-     * @return \Illuminate\Contracts\View\Factory|\Illuminate\View\View
-     */
-    public function sectionStudents($section_id)
-    {
-        $students = $this->userService->getSectionStudentsWithSchool($section_id);
-
-        return view('profile.section-students', compact('students'));
-    }
-
-    /**
-     * @param $section_id
-     * @return \Illuminate\Contracts\View\Factory|\Illuminate\View\View
-     */
-    public function promoteSectionStudents(Request $request, $section_id)
-    {
-        if($this->userService->hasSectionId($section_id))
-            return $this->userService->promoteSectionStudentsView(
-                $this->userService->getSectionStudentsWithStudentInfo($request, $section_id),
-                Myclass::with('sections')->bySchool(\Auth::user()->school_id)->get(),
-                $section_id
-            );
-        else
-            return $this->userService->promoteSectionStudentsView([], [], $section_id);
-    }
-
-    /**
-     * @param Request $request
-     * @return \Illuminate\Http\RedirectResponse
-     */
-    public function promoteSectionStudentsPost(Request $request)
-    {   
-        return $this->userService->promoteSectionStudentsPost($request);
-    }
-
-    /**
-     * @return \Illuminate\Contracts\View\Factory|\Illuminate\View\View
-     */
-    public function changePasswordGet()
-    {
-        return view('profile.change-password');
-    }
-
-    /**
-     * @param ChangePasswordRequest $request
-     * @return \Illuminate\Http\RedirectResponse
-     */
-    public function changePasswordPost(ChangePasswordRequest $request)
-    {
-        if (Hash::check($request->old_password, Auth::user()->password)) {
-            $request->user()->fill([
-              'password' => Hash::make($request->new_password),
-            ])->save();
-
-            return back()->with('status', __('Saved'));
-        }
-
-        return back()->with('error-status', __('Passwords do not match.'));
-    }
-
-    /**
-     * @return \Illuminate\Contracts\View\Factory|\Illuminate\View\View
-     */
-    public function impersonateGet()
-    {
-        if (app('impersonate')->isImpersonating()) {
-            Auth::user()->leaveImpersonation();
-            return (Auth::user()->role == 'master')?redirect('/masters') : redirect('/home');
-        }
-        else {
-            return view('profile.impersonate', [
-                'other_users' => $this->user->where('id', '!=', auth()->id())->get([ 'id', 'name', 'role' ])
-            ]);
+            return back()->with('status', 'Teacher creation was successful!');
+        } catch (\Exception $e) {
+            return back()->withError($e->getMessage());
         }
     }
 
-    /**
-     * @return \Illuminate\Contracts\View\Factory|\Illuminate\View\View
-     */
-    public function impersonate(ImpersonateUserRequest $request)
-    {
-        $user = $this->user->find($request->id);
-        Auth::user()->impersonate($user);
-        return redirect('/home');
+    public function getStudentList(Request $request) {
+        $current_school_session_id = $this->getSchoolCurrentSession();
+
+        $class_id = $request->query('class_id', 0);
+        $section_id = $request->query('section_id', 0);
+
+        try{
+
+            $school_classes = $this->schoolClassRepository->getAllBySession($current_school_session_id);
+
+            $studentList = $this->userRepository->getAllStudents($current_school_session_id, $class_id, $section_id);
+
+            $data = [
+                'studentList'       => $studentList,
+                'school_classes'    => $school_classes,
+            ];
+
+            return view('students.list', $data);
+        } catch (\Exception $e) {
+            return back()->withError($e->getMessage());
+        }
+    }
+
+
+    public function showStudentProfile($id) {
+        $student = $this->userRepository->findStudent($id);
+
+        $current_school_session_id = $this->getSchoolCurrentSession();
+        $promotionRepository = new PromotionRepository();
+        $promotion_info = $promotionRepository->getPromotionInfoById($current_school_session_id, $id);
+
+        $data = [
+            'student'           => $student,
+            'promotion_info'    => $promotion_info,
+        ];
+
+        return view('students.profile', $data);
+    }
+
+    public function showTeacherProfile($id) {
+        $teacher = $this->userRepository->findTeacher($id);
+        $data = [
+            'teacher'   => $teacher,
+        ];
+        return view('teachers.profile', $data);
+    }
+
+
+    public function createStudent() {
+        $current_school_session_id = $this->getSchoolCurrentSession();
+
+        $school_classes = $this->schoolClassRepository->getAllBySession($current_school_session_id);
+
+        $data = [
+            'current_school_session_id' => $current_school_session_id,
+            'school_classes'            => $school_classes,
+        ];
+
+        return view('students.add', $data);
     }
 
     /**
      * Store a newly created resource in storage.
      *
-     * @param CreateUserRequest $request
-     *
+     * @param  StudentStoreRequest $request
      * @return \Illuminate\Http\Response
      */
-    public function store(CreateUserRequest $request)
+    public function storeStudent(StudentStoreRequest $request)
     {
-        DB::transaction(function () use ($request) {
-            $password = $request->password;
-            $tb = $this->userService->storeStudent($request);
-            try {
-                // Fire event to store Student information
-                if(event(new StudentInfoUpdateRequested($request,$tb->id))){
-                    // Fire event to send welcome email
-                    event(new UserRegistered($tb, $password));
-                } else {
-                    throw new \Exeception('Event returned false');
-                }
-            } catch(\Exception $ex) {
-                Log::info('Email failed to send to this address: '.$tb->email.'\n'.$ex->getMessage());
-            }
-        });
-
-        return back()->with('status', __('Saved'));
-    }
-
-    /**
-     * @param CreateAdminRequest $request
-     * @return \Illuminate\Http\RedirectResponse
-     */
-    public function storeAdmin(CreateAdminRequest $request)
-    {
-        $password = $request->password;
-        $tb = $this->userService->storeAdmin($request);
         try {
-            // Fire event to send welcome email
-            // event(new userRegistered($userObject, $plain_password)); // $plain_password(optional)
-            event(new UserRegistered($tb, $password));
-        } catch(\Exception $ex) {
-            Log::info('Email failed to send to this address: '.$tb->email);
-        }
+            $this->userRepository->createStudent($request->validated());
 
-        return back()->with('status', __('Saved'));
+            return back()->with('status', 'Student creation was successful!');
+        } catch (\Exception $e) {
+            return back()->withError($e->getMessage());
+        }
     }
 
-    /**
-     * @param CreateTeacherRequest $request
-     * @return \Illuminate\Http\RedirectResponse
-     */
-    public function storeTeacher(CreateTeacherRequest $request)
-    {
-        $password = $request->password;
-        $tb = $this->userService->storeStaff($request, 'teacher');
+    public function editStudent($student_id) {
+        $student = $this->userRepository->findStudent($student_id);
+        $studentParentInfoRepository = new StudentParentInfoRepository();
+        $parent_info = $studentParentInfoRepository->getParentInfo($student_id);
+        $promotionRepository = new PromotionRepository();
+        $current_school_session_id = $this->getSchoolCurrentSession();
+        $promotion_info = $promotionRepository->getPromotionInfoById($current_school_session_id, $student_id);
+
+        $data = [
+            'student'       => $student,
+            'parent_info'   => $parent_info,
+            'promotion_info'=> $promotion_info,
+        ];
+        return view('students.edit', $data);
+    }
+
+    public function updateStudent(Request $request) {
         try {
-            // Fire event to send welcome email
-            event(new UserRegistered($tb, $password));
-        } catch(\Exception $ex) {
-            Log::info('Email failed to send to this address: '.$tb->email);
-        }
+            $this->userRepository->updateStudent($request->toArray());
 
-        return back()->with('status', __('Saved'));
+            return back()->with('status', 'Student update was successful!');
+        } catch (\Exception $e) {
+            return back()->withError($e->getMessage());
+        }
     }
 
-    /**
-     * @param CreateAccountantRequest $request
-     * @return \Illuminate\Http\RedirectResponse
-     */
-    public function storeAccountant(CreateAccountantRequest $request)
-    {
-        $password = $request->password;
-        $tb = $this->userService->storeStaff($request, 'accountant');
+    public function editTeacher($teacher_id) {
+        $teacher = $this->userRepository->findTeacher($teacher_id);
+
+        $data = [
+            'teacher'   => $teacher,
+        ];
+
+        return view('teachers.edit', $data);
+    }
+    public function updateTeacher(Request $request) {
         try {
-            // Fire event to send welcome email
-            event(new UserRegistered($tb, $password));
-        } catch(\Exception $ex) {
-            Log::info('Email failed to send to this address: '.$tb->email);
+            $this->userRepository->updateTeacher($request->toArray());
+
+            return back()->with('status', 'Teacher update was successful!');
+        } catch (\Exception $e) {
+            return back()->withError($e->getMessage());
         }
-
-        return back()->with('status', __('Saved'));
     }
 
-    /**
-     * @param CreateLibrarianRequest $request
-     * @return \Illuminate\Http\RedirectResponse
-     */
-    public function storeLibrarian(CreateLibrarianRequest $request)
-    {
-        $password = $request->password;
-        $tb = $this->userService->storeStaff($request, 'librarian');
-        try {
-            // Fire event to send welcome email
-            event(new UserRegistered($tb, $password));
-        } catch(\Exception $ex) {
-            Log::info('Email failed to send to this address: '.$tb->email);
-        }
+    public function getTeacherList(){
+        $teachers = $this->userRepository->getAllTeachers();
 
-        return back()->with('status', __('Saved'));
-    }
+        $data = [
+            'teachers' => $teachers,
+        ];
 
-    /**
-     * Display the specified resource.
-     *
-     * @param int $id
-     *
-     * @return UserResource
-     */
-    public function show($user_code)
-    {
-        $user = $this->userService->getUserByUserCode($user_code);
-
-        return view('profile.user', compact('user'));
-    }
-
-    /**
-     * Show the form for editing the specified resource.
-     *
-     * @param int $id
-     *
-     * @return \Illuminate\Http\Response
-     */
-    public function edit($id)
-    {
-        $user = $this->user->find($id);
-        
-        $classes = Myclass::query()
-            ->bySchool($user->school_id)
-            ->pluck('id')
-            ->toArray();
-
-        $sections = Section::query()
-            ->whereIn('class_id', $classes)
-            ->get();
-
-        $departments = Department::query()
-            ->bySchool($user->school_id)
-            ->get();
-
-        return view('profile.edit', [
-            'user' => $user,
-            'sections' => $sections,
-            'departments' => $departments,
-        ]);
-    }
-
-    /**
-     * Update the specified resource in storage.
-     *
-     * @param UpdateUserRequest $request
-     * @return \Illuminate\Http\Response
-     */
-    public function update(UpdateUserRequest $request)
-    {
-		 
-        DB::transaction(function () use ($request) {
-            $tb = $this->user->find($request->user_id);
-            $tb->name = $request->name;
-            $tb->email = (!empty($request->email)) ? $request->email : '';
-            $tb->nationality = (!empty($request->nationality)) ? $request->nationality : '';
-            $tb->phone_number = $request->phone_number;
-            $tb->address = (!empty($request->address)) ? $request->address : '';
-            $tb->about = (!empty($request->about)) ? $request->about : '';
-			if (!empty($request->pic_path)) {
-				$tb->pic_path = $request->pic_path;
-			}
-            if ($request->user_role == 'teacher') {
-                $tb->department_id = $request->department_id;
-                $tb->section_id = $request->class_teacher_section_id;
-            }
-            if ($tb->save()) {
-                if ($request->user_role == 'student') {
-                    // $request->validate([
-                    //   'session' => 'required',
-                    //   'version' => 'required',
-                    //   'birthday' => 'required',
-                    //   'religion' => 'required',
-                    //   'father_name' => 'required',
-                    //   'mother_name' => 'required',
-                    // ]);
-                    try{
-                        // Fire event to store Student information
-                        event(new StudentInfoUpdateRequested($request,$tb->id));
-                    } catch(\Exception $ex) {
-                        Log::info('Failed to update Student information, Id: '.$tb->id. 'err:'.$ex->getMessage());
-                    }
-                }
-            }
-        });
-
-        return back()->with('status', __('Saved'));
-    }
-
-    /**
-     * Activate admin
-     * @param $id
-     * @return \Illuminate\Http\RedirectResponse
-     */
-    public function activateAdmin($id)
-    {
-        $admin = $this->user->find($id);
-
-        if ($admin->active !== 0) {
-            $admin->active = 0;
-        } else {
-            $admin->active = 1;
-        }
-
-        $admin->save();
-
-        return back()->with('status', __('Saved'));
-    }
-
-    /**
-     * Deactivate admin
-     * @param $id
-     * @return \Illuminate\Http\RedirectResponse
-     */
-    public function deactivateAdmin($id)
-    {
-       $admin = $this->user->find($id);
-
-        if ($admin->active !== 1) {
-            $admin->active = 1;
-        } else {
-            $admin->active = 0;
-        }
-
-        $admin->save();
-
-        return back()->with('status', __('Saved'));
-    }
-
-    /**
-     * Remove the specified resource from storage.
-     *
-     * @param int $id
-     *
-     * @return void
-     */
-    public function destroy($id)
-    {
-        // return ($this->user->destroy($id))?response()->json([
-      //   'status' => 'success'
-      // ]):response()->json([
-      //   'status' => 'error'
-      // ]);
+        return view('teachers.list', $data);
     }
 }
