@@ -1,44 +1,109 @@
-<!DOCTYPE html>
-<html lang="{{ app()->getLocale() }}">
-
+<!doctype html>
+<html lang="{{ str_replace('_', '-', app()->getLocale()) }}">
 <head>
     <meta charset="utf-8">
-    <meta http-equiv="X-UA-Compatible" content="IE=edge">
     <meta name="viewport" content="width=device-width, initial-scale=1">
 
     <!-- CSRF Token -->
     <meta name="csrf-token" content="{{ csrf_token() }}">
 
-    <title>@yield('title') - {{ (Auth::check() && (Auth::user()->role == 'student' || Auth::user()->role == 'teacher'
-        || Auth::user()->role == 'admin' || Auth::user()->role == 'accountant' || Auth::user()->role ==
-        'librarian'))?Auth::user()->school->name:config('app.name') }}</title>
+    <title>{{ config('app.name', 'Unifiedtransform') }}</title>
 
-    <link rel="stylesheet" href="{{ url('css/loader.css') }}">
+    <link rel="shortcut icon" href="{{asset('favicon_io/favicon.ico')}}">
+    <link rel="shortcut icon" sizes="16x16" href="{{asset('favicon_io/favicon-16x16.png')}}">
+    <link rel="shortcut icon" sizes="32x32" href="{{asset('favicon_io/favicon-32x32.png')}}">
+    <link rel="apple-touch-icon" href="{{asset('favicon_io/apple-touch-icon.png')}}">
+    <link rel="icon" href="{{asset('favicon_io/android-chrome-192x192.png')}}" sizes="192x192">
+    <link rel="icon" href="{{asset('favicon_io/android-chrome-512x512.png')}}" sizes="512x512">
 
-    <script src="{{ url('js/vendors.js') }}"></script>
+    <!-- Scripts -->
+    <script src="{{ asset('js/bootstrap.bundle.min.js') }}"></script>
+    <!-- <script src="{{ asset('js/app.js') }}" defer></script> -->
 
-    <script type="text/javascript">
-        $.ajaxSetup({
-            headers: {
-                'X-CSRF-TOKEN': $('meta[name="csrf-token"]').attr('content')
-            }
-        });
-    </script>
-    <script src="{{ url('js/application.js') }}"></script>
-    @yield('after_scripts')
+    <!-- Fonts -->
+    <link rel="dns-prefetch" href="//fonts.gstatic.com">
+    <link href="https://fonts.googleapis.com/css?family=Nunito" rel="stylesheet">
+
+   
+    <link href="{{ asset('css/bootstrap.min.css') }}" rel="stylesheet">
+    <link href="{{ asset('css/app.css') }}" rel="stylesheet">
 </head>
-
 <body>
-    @include('components.loader')
     <div id="app">
-        @include('components.navbar-top')
-        @yield('content')
+        <nav class="navbar sticky-top navbar-expand-md navbar-light bg-white border-btm-e6">
+            <div class="container">
+                <a class="navbar-brand" href="{{ url('/') }}">
+                    <i class="bi bi-house"></i> {{ config('app.name', 'Laravel') }}
+                </a>
+                <button class="navbar-toggler" type="button" data-bs-toggle="collapse" data-bs-target="#navbarSupportedContent" aria-controls="navbarSupportedContent" aria-expanded="false" aria-label="{{ __('Toggle navigation') }}">
+                    <span class="navbar-toggler-icon"></span>
+                </button>
+
+                <div class="collapse navbar-collapse" id="navbarSupportedContent">
+                    <!-- Left Side Of Navbar -->
+                    @auth
+                        @php
+                            $latest_school_session = \App\Models\SchoolSession::latest()->first();
+                            $current_school_session_name = null;
+                            if($latest_school_session){
+                                $current_school_session_name = $latest_school_session->session_name;
+                            }
+                        @endphp
+                    <ul class="navbar-nav">
+                        <li class="nav-item">
+                            @if (session()->has('browse_session_name') && session('browse_session_name') !== $current_school_session_name)
+                                <a class="nav-link text-danger disabled" href="#" tabindex="-1" aria-disabled="true"><i class="bi bi-exclamation-diamond-fill me-2"></i> Browsing as Academic Session {{session('browse_session_name')}}</a>
+                            @elseif(\App\Models\SchoolSession::latest()->count() > 0)
+                                <a class="nav-link disabled" href="#" tabindex="-1" aria-disabled="true">Current Academic Session {{$current_school_session_name}}</a>
+                            @else
+                                <a class="nav-link text-danger disabled" href="#" tabindex="-1" aria-disabled="true"><i class="bi bi-exclamation-diamond-fill me-2"></i> Create an Academic Session.</a>
+                            @endif
+                        </li>
+                    </ul>
+                    @endauth
+                    <!-- Right Side Of Navbar -->
+                    <ul class="navbar-nav ms-auto">
+                        <!-- Authentication Links -->
+                        @guest
+                            @if (Route::has('login'))
+                                <li class="nav-item">
+                                    <a class="nav-link" href="{{ route('login') }}">{{ __('Login') }}</a>
+                                </li>
+                            @endif
+                        @else
+                            <li class="nav-item dropdown">
+                                <a id="navbarDropdown" class="nav-link dropdown-toggle" href="#" role="button" data-bs-toggle="dropdown" aria-haspopup="true" aria-expanded="false">
+                                    <span class="badge bg-light text-dark">{{ ucfirst(Auth::user()->role) }}</span>
+                                    {{ Auth::user()->first_name }} {{ Auth::user()->last_name }}
+                                </a>
+
+                                <div class="dropdown-menu dropdown-menu-end" aria-labelledby="navbarDropdown">
+                                    <a class="dropdown-item" href="{{route('password.edit')}}">
+                                        <i class="bi bi-key me-2"></i> Change Password
+                                    </a>
+                                    <a class="dropdown-item" href="{{ route('logout') }}"
+                                       onclick="event.preventDefault();
+                                                     document.getElementById('logout-form').submit();">
+                                        <i class="bi bi-door-open me-2"></i> {{ __('Logout') }}
+                                    </a>
+
+                                    <form id="logout-form" action="{{ route('logout') }}" method="POST" class="d-none">
+                                        @csrf
+                                    </form>
+                                </div>
+                            </li>
+                        @endguest
+                    </ul>
+                </div>
+            </div>
+        </nav>
+        <main>
+            @yield('content')
+        </main>
     </div>
 
-    <link href="https://fonts.googleapis.com/icon?family=Material+Icons&style=normal&weight=400"
-      rel="stylesheet">
-    <link rel="stylesheet" href="{{ url('css/vendors.css') }}" id="bootswatch-print-id">
-    <link rel="stylesheet" href="{{ url('css/application.css') }}">
+    <div id="watermark">
+        <p>Unifiedtransform</p>
+    </div>
 </body>
-
 </html>
